backend/core/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.11
python-dotenv
sqlparse==0.5.3
tzdata==2025.2
